Updated level generation and parallaxing
The level generator script now creates a child object named after it's
layerindex (which represents the z-axis) and spawns objects as children
to this layerobject. This separation of layers makes level generation
much easier. I also borrowed a camera script from the github project
linked in this article:
http://www.gamasutra.com/blogs/MichalBerlinger/20160323/268657/Combining_Perspective_and_Orthographic_Camera_for_Parallax_Effect_in_2D_Game.php

This script basically helps with setting up the correct values for the
three cameras in order to have a nice looking parallax effect.